Moved 5 error tests to new test_api_errors.py
